refactor: 출력 뷰에서 메시지 생성부분 전용 클래스 생성
